list

+++ build.sh
@@ -8,7 +8,13 @@ RENDER_WORKING_DIR=support
 
 # Cambia al directorio de trabajo de Render
 cd $RENDER_WORKING_DIR
-pip install -r requirements.txt
+# Instala las dependencias desde requirements.txt si existe
+if [ -f "requirements.txt" ]; then
+    pip install -r requirements.txt
+else
+    echo "El archivo requirements.txt no se encontró en la ubicación especificada."
+    exit 1
+fi
 
 python manage.py collectstatic --no-input
 python manage.py migrate
